+++ src/sections/dashboard/view.js
@@ -15,161 +15,9 @@ import { useSettingsContext } from 'src/components/settings';
 import { useAuthContext } from 'src/auth/hooks';
 import AnalyticsCurrentVisits from './analytics-current-visits';
 import AnalyticsCard from './analytics-card';
+import {handleValue} from './func';
 
-// ----------------------------------------------------------------------
-
-const checkData = (totalResult, temp, item, coaData) => {
-  const tempResult = totalResult
-  const setTotalResult = (type) => {
-    let flag = false;
-    if (type) {
-      if(tempResult.revenue.length > 0) {
-        tempResult.revenue.forEach((revenue) => {
-          if (revenue.name === item[3]) {
-            revenue.value += item[2];
-            flag = true;
-          }
-        });
-      } 
-      if (flag === false) tempResult.revenue.push({ name: item[3], value: item[2] });
-    } else {
-      if(tempResult.expense.length > 0) {
-        tempResult.expense.forEach((expense) => {
-          if (expense.name === item[3]) {
-            expense.value += item[2];
-            flag = true;
-          }
-        });
-      }
-      if (flag === false) tempResult.expense.push({ name: item[3], value: item[2] });
-    }
-  };
-
-  if (coaData[0].toString() === item[0]) {
-    switch (coaData[2]) {
-      case 'Revenues':
-        temp.revenue += item[2];
-        totalResult = setTotalResult(true);
-        break;
-      case 'COGS':
-        temp.cogs += item[2];
-        totalResult = setTotalResult(false);
-        break;
-      case 'S&O':
-        temp.so += item[2];
-        totalResult = setTotalResult(false);
-        break;
-      case 'SGA':
-        temp.sga += item[2];
-        totalResult = setTotalResult(false);
-        break;
-      case 'Salaries & Benefits':
-        temp.salaries += item[2];
-        totalResult = setTotalResult(false);
-        break;
-      case 'Tax':
-        temp.taxes += item[2];
-        totalResult = setTotalResult(false);
-        break;
-      case 'Amortization & Depr':
-        temp.amorization += item[2];
-        totalResult = setTotalResult(false);
-        break;
-      case 'Projects':
-        temp.project += item[2];
-        totalResult = setTotalResult(false);
-        break;
-      default:
-        break;
-    }
-  }
-};
-
-// monthly data
-const divideData = (mData, coa) => {
-  const totalResult = { revenue: [], expense: []};
-  const temp = { revenue: 0, cogs: 0, so: 0, sga: 0, salaries: 0, taxes: 0, amorization: 0, project: 0 };
-  mData.forEach((item) => {
-    coa.map( (coaData) => {
-      if(coaData[0].toString()===item[0]) {
-        checkData(totalResult, temp, item, coaData)
-      } 
-      return true 
-    });
-  });
-  const resultData = {
-    monthly: totalResult, eachItem: temp
-  }
-  return resultData;
-};
-
-// total result
-const handleValue = (data,coa) => {
-  const totalResult = {Revenue:[], Expense:[], TotalRevenue:[],NetIncome:[],GrossProfit:[]}
-
-  const checkResult = (tempData, flag) =>{
-    tempData.forEach(item => {
-      let exist = true;
-      if(flag === true) {
-        totalResult.Revenue.map((revenue => {
-          if(item.name === revenue.name){
-            revenue.value += item.value;
-            exist = false;
-          }
-          return true;
-        }))
-        if(exist === true) totalResult.Revenue.push(item)
-      } else {
-        totalResult.Expense.map(expense => {
-          if(item.name === expense.name){
-            expense.value += item.value;
-            exist = false;
-          }
-          return true;
-        })  
-        if(exist === true) totalResult.Revenue.push(item)
-      }
-    })
-  }
-
-  const addMonthlyResult = (monthlyResult, date) => {
-    const {revenue, amorization, cogs, project, salaries, sga, so, taxes} = monthlyResult.eachItem
-    
-    const totalRevenue = revenue; 
-    const netIncome= revenue - (so + sga+ salaries + taxes + amorization + project);
-    const grossProfit= revenue - (cogs + so);
-    totalResult.TotalRevenue.push({year:date, value:totalRevenue})
-    totalResult.NetIncome.push({year:date, value:netIncome})
-    totalResult.GrossProfit.push({year:date, value:grossProfit})
-    
-    if(totalResult.Revenue.length === 0) {
-      totalResult.Revenue.push(...monthlyResult.monthly.revenue)
-    } else {
-      checkResult(monthlyResult.monthly.revenue, true) 
-    }
-
-    if(totalResult.Expense.length === 0) {
-      totalResult.Expense.push(...monthlyResult.monthly.expense)
-    } else {
-      checkResult(monthlyResult.monthly.expense, false)
-    }
-  }
-
-  data.forEach((item) => {
-    if (item.data.length > 0 && Array.isArray(item.data)) {
-      const date = item.dateFlag;
-      const temp = [...item.data];
-      // item.data => one month's result => It's array.
-      const monthlyResult = divideData(temp, coa);
-      addMonthlyResult(monthlyResult, date);
-    } else {
-      console.log('empty')
-    }
-  });
-  return totalResult
-};
-
-export default function Dashboard() {
+export default function IncomeStatement() {
   const settings = useSettingsContext();
   const [result, setResult] = useState({
     Revenue:null, Expense:null, TotalRevenue:[], NetIncome:[], GrossProfit:[]
